refactor navbar menu button
